ml-model/requirements.txt
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
tensorboard==2.15.1
albumentations==1.3.1
imgaug==0.4.0
